+++ packages/calendar/src/features/month-picker/MonthPicker.stories.tsx
@@ -1,6 +1,6 @@
 import { Store, withState } from "@dump247/storybook-state";
 import * as React from "react";
-import { Months } from "@stenajs-webui/calendar/dist/util/calendar/CalendarDataFactory";
+import { Months } from "@stenajs-webui/calendar";
 import { MonthPicker } from "@stenajs-webui/calendar";
 
 interface DateRangeState {