+++ packages/calendar/src/features/month-picker/MonthPicker.tsx
@@ -1,16 +1,16 @@
 import * as React from "react";
-import { Months } from "../../util/calendar/CalendarDataFactory";
+import { Month } from "../../util/calendar/CalendarDataFactory";
 import { ValueAndOnValueChangeProps } from "@stenajs-webui/forms";
 import { Column, Row } from "@stenajs-webui/core";
 import { MonthPickerCell } from "./MonthPickerCell";
 
-interface Props extends ValueAndOnValueChangeProps<Months> {}
+interface Props extends ValueAndOnValueChangeProps<Month> {}
 
 const monthMatrix = [
-  [Months.JANUARY, Months.FEBRUARY, Months.MARCH],
-  [Months.APRIL, Months.MAY, Months.JUNE],
-  [Months.JULY, Months.AUGUST, Months.SEPTEMBER],
-  [Months.OCTOBER, Months.NOVEMBER, Months.DECEMBER],
+  [Month.JANUARY, Month.FEBRUARY, Month.MARCH],
+  [Month.APRIL, Month.MAY, Month.JUNE],
+  [Month.JULY, Month.AUGUST, Month.SEPTEMBER],
+  [Month.OCTOBER, Month.NOVEMBER, Month.DECEMBER],
 ];
 
 export const MonthPicker: React.FC<Props> = ({ value, onValueChange }) => {